refactor: update imports and remove backward compatibility
- Update __init__.py to export new modular structure
- Remove deprecated util.py in favor of utils module
- Update logic modules to use centralized utilities
- Clean up import paths across modules
