+++ src/property_driven_ml/__init__.py
@@ -15,9 +15,6 @@ from . import logics
 from . import constraints
 from . import training
 
-# Import utilities directly for convenience
-from .util import safe_div, safe_pow, maybe
-
 __version__ = "0.1.0"
 __author__ = "Thomas Flinkow"
 
@@ -25,8 +22,5 @@ __all__ = [
     # Submodules
     "logics",
     "constraints", 
-    "training",
-    
-    # Utilities
-    "safe_div", "safe_pow", "maybe"
+    "training"
 ]
