+++ src/property_driven_ml/logics/fuzzy_logics.py
@@ -2,7 +2,7 @@ import torch
 
 from .logic import Logic
 
-from ..util import safe_div, safe_pow
+from ..utils import safe_div, safe_pow
 
 class FuzzyLogic(Logic):
     def __init__(self, name: str):
